pyproject.toml
[tool.poetry]
name = "CodeDigest"
version = "0.1.0"
description = "Summarize projects using LLMs"
authors = ["Carlo Kroll <ckroll95@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
openai = "^0.14.0"
pydantic = "^1.8.2"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
